# contracts/openapi.json
{
  "components": {
    "schemas": {
      "AnalysisResult": {
        "additionalProperties": false,
        "properties": {
          "cause_candidates": {
            "items": {
              "type": "string"
            },
            "title": "Cause Candidates",
            "type": "array"
          },
          "confidence": {
            "anyOf": [
              {
                "maximum": 1.0,
                "minimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Confidence"
          },
          "evidence": {
            "items": {
              "type": "string"
            },
            "title": "Evidence",
            "type": "array"
          },
          "inspection_id": {
            "format": "uuid",
            "title": "Inspection Id",
            "type": "string"
          },
          "is_mock_input": {
            "title": "Is Mock Input",
            "type": "boolean"
          },
          "model_mode": {
            "anyOf": [
              {
                "enum": [
                  "rgb",
                  "sensor",
                  "fusion"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model Mode"
          },
          "model_version": {
            "anyOf": [
              {
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model Version"
          },
          "origin": {
            "enum": [
              "none",
              "mock",
              "real_model"
            ],
            "title": "Origin",
            "type": "string"
          },
          "outcome": {
            "anyOf": [
              {
                "enum": [
                  "no_visible_abnormality",
                  "suspected_abnormality",
                  "inconclusive"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Outcome"
          },
          "schema_version": {
            "const": "0.1.0",
            "title": "Schema Version",
            "type": "string"
          },
          "status": {
            "enum": [
              "pending_model",
              "completed",
              "failed"
            ],
            "title": "Status",
            "type": "string"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "title": "Warnings",
            "type": "array"
          }
        },
        "required": [
          "schema_version",
          "inspection_id",
          "status",
          "origin",
          "is_mock_input",
          "outcome",
          "model_version",
          "model_mode",
          "confidence",
          "evidence",
          "cause_candidates",
          "warnings"
        ],
        "title": "AnalysisResult",
        "type": "object"
      },
      "Body_upload_inspection_v1_inspections_post": {
        "properties": {
          "metadata": {
            "title": "Metadata",
            "type": "string"
          },
          "photo": {
            "contentMediaType": "application/octet-stream",
            "title": "Photo",
            "type": "string"
          }
        },
        "required": [
          "photo",
          "metadata"
        ],
        "title": "Body_upload_inspection_v1_inspections_post",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "InspectionMetadata": {
        "additionalProperties": false,
        "properties": {
          "batch_id": {
            "maxLength": 80,
            "minLength": 1,
            "pattern": "^[A-Za-z0-9_-]+$",
            "title": "Batch Id",
            "type": "string"
          },
          "capture_source": {
            "enum": [
              "real",
              "synthetic"
            ],
            "title": "Capture Source",
            "type": "string"
          },
          "captured_at": {
            "format": "date-time",
            "title": "Captured At",
            "type": "string"
          },
          "crop_code": {
            "enum": [
              "cherry_tomato",
              "pepper",
              "cucumber",
              "lettuce",
              "bok_choy",
              "eggplant",
              "basil",
              "rosemary"
            ],
            "title": "Crop Code",
            "type": "string"
          },
          "image_media_type": {
            "enum": [
              "image/jpeg",
              "image/png"
            ],
            "title": "Image Media Type",
            "type": "string"
          },
          "image_sha256": {
            "pattern": "^[0-9a-f]{64}$",
            "title": "Image Sha256",
            "type": "string"
          },
          "inspection_id": {
            "format": "uuid",
            "title": "Inspection Id",
            "type": "string"
          },
          "parent_plant_id": {
            "anyOf": [
              {
                "maxLength": 80,
                "minLength": 1,
                "pattern": "^[A-Za-z0-9_-]+$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent Plant Id"
          },
          "schema_version": {
            "const": "0.1.0",
            "title": "Schema Version",
            "type": "string"
          },
          "sensor_absence_reason": {
            "anyOf": [
              {
                "enum": [
                  "not_acquired",
                  "disconnected",
                  "permission_denied",
                  "read_error"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sensor Absence Reason"
          },
          "sensor_packet": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SensorPacket"
              },
              {
                "type": "null"
              }
            ]
          },
          "sensor_received_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sensor Received At"
          },
          "subject_id": {
            "maxLength": 80,
            "minLength": 1,
            "pattern": "^[A-Za-z0-9_-]+$",
            "title": "Subject Id",
            "type": "string"
          },
          "subject_type": {
            "enum": [
              "plant",
              "fruit"
            ],
            "title": "Subject Type",
            "type": "string"
          }
        },
        "required": [
          "schema_version",
          "inspection_id",
          "subject_id",
          "subject_type",
          "parent_plant_id",
          "batch_id",
          "crop_code",
          "captured_at",
          "capture_source",
          "image_sha256",
          "image_media_type",
          "sensor_packet",
          "sensor_received_at",
          "sensor_absence_reason"
        ],
        "title": "InspectionMetadata",
        "type": "object"
      },
      "SensorPacket": {
        "additionalProperties": false,
        "properties": {
          "boot_id": {
            "maxLength": 80,
            "minLength": 1,
            "pattern": "^[A-Za-z0-9_-]+$",
            "title": "Boot Id",
            "type": "string"
          },
          "calibration_id": {
            "anyOf": [
              {
                "maxLength": 80,
                "minLength": 1,
                "pattern": "^[A-Za-z0-9_-]+$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Calibration Id"
          },
          "device_id": {
            "maxLength": 80,
            "minLength": 1,
            "pattern": "^[A-Za-z0-9_-]+$",
            "title": "Device Id",
            "type": "string"
          },
          "errors": {
            "items": {
              "maxLength": 80,
              "minLength": 1,
              "type": "string"
            },
            "title": "Errors",
            "type": "array"
          },
          "firmware_version": {
            "maxLength": 40,
            "minLength": 1,
            "title": "Firmware Version",
            "type": "string"
          },
          "message_type": {
            "const": "measurement",
            "title": "Message Type",
            "type": "string"
          },
          "mode": {
            "enum": [
              "real",
              "mock"
            ],
            "title": "Mode",
            "type": "string"
          },
          "readings": {
            "$ref": "#/components/schemas/SensorReadings"
          },
          "request_id": {
            "format": "uuid",
            "title": "Request Id",
            "type": "string"
          },
          "schema_version": {
            "const": "0.1.0",
            "title": "Schema Version",
            "type": "string"
          },
          "sequence": {
            "minimum": 0.0,
            "title": "Sequence",
            "type": "integer"
          },
          "uptime_ms": {
            "minimum": 0.0,
            "title": "Uptime Ms",
            "type": "integer"
          }
        },
        "required": [
          "schema_version",
          "message_type",
          "request_id",
          "device_id",
          "boot_id",
          "firmware_version",
          "sequence",
          "uptime_ms",
          "mode",
          "calibration_id",
          "readings",
          "errors"
        ],
        "title": "SensorPacket",
        "type": "object"
      },
      "SensorReadings": {
        "additionalProperties": false,
        "properties": {
          "air_rh_pct": {
            "anyOf": [
              {
                "maximum": 100.0,
                "minimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Air Rh Pct"
          },
          "air_temp_c": {
            "anyOf": [
              {
                "minimum": -273.15,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Air Temp C"
          },
          "soil_adc_bits": {
            "enum": [
              10,
              12,
              14
            ],
            "title": "Soil Adc Bits",
            "type": "integer"
          },
          "soil_index": {
            "anyOf": [
              {
                "maximum": 1.0,
                "minimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Soil Index"
          },
          "soil_raw": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Soil Raw"
          }
        },
        "required": [
          "soil_adc_bits",
          "soil_raw",
          "soil_index",
          "air_temp_c",
          "air_rh_pct"
        ],
        "title": "SensorReadings",
        "type": "object"
      },
      "StoredInspection": {
        "properties": {
          "analysis": {
            "$ref": "#/components/schemas/AnalysisResult"
          },
          "inspection_id": {
            "format": "uuid",
            "title": "Inspection Id",
            "type": "string"
          },
          "metadata": {
            "$ref": "#/components/schemas/InspectionMetadata"
          },
          "received_at_utc": {
            "format": "date-time",
            "title": "Received At Utc",
            "type": "string"
          }
        },
        "required": [
          "inspection_id",
          "metadata",
          "analysis",
          "received_at_utc"
        ],
        "title": "StoredInspection",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Local development storage; no authentication or diagnostic model.",
    "title": "VisionMachine Foundation",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/healthz": {
      "get": {
        "operationId": "health_healthz_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Health"
      }
    },
    "/v1/capabilities": {
      "get": {
        "operationId": "capabilities_v1_capabilities_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Capabilities"
      }
    },
    "/v1/dev/validate-inspection": {
      "post": {
        "operationId": "validate_inspection_v1_dev_validate_inspection_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InspectionMetadata"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalysisResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Validate Inspection"
      }
    },
    "/v1/inspections": {
      "post": {
        "operationId": "upload_inspection_v1_inspections_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_inspection_v1_inspections_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StoredInspection"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Upload Inspection",
        "tags": [
          "local inspections"
        ]
      }
    },
    "/v1/inspections/{inspection_id}": {
      "get": {
        "operationId": "get_inspection_v1_inspections__inspection_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "inspection_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Inspection Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StoredInspection"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Inspection",
        "tags": [
          "local inspections"
        ]
      }
    }
  }
}